moved navbar, header & landingPage to /components; started api.config functionality
