+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@types/react": "^16.9.56",
+    "dotenv": "^8.2.0",
     "react": "16.13.1",
     "react-hook-form": "^6.12.2",
     "react-native": "0.63.3",
@@ -21,6 +22,7 @@
     "@babel/core": "^7.8.4",
     "@babel/runtime": "^7.8.4",
     "@react-native-community/eslint-config": "^1.1.0",
+    "@types/dotenv": "^8.2.0",
     "@types/jest": "^25.2.3",
     "@types/react-native": "^0.63.2",
     "@types/react-test-renderer": "^16.9.2",
@@ -31,6 +33,7 @@
     "jest": "^25.1.0",
     "metro-react-native-babel-preset": "^0.59.0",
     "react-test-renderer": "16.13.1",
+    "ts-node": "^9.1.1",
     "typescript": "^3.8.3"
   },
   "jest": {
